refactored folder structure for home component
